Update pollutantmean.R
Given data (specdata)  contains 332 comma-separated-value (CSV) files containing pollution monitoring data for fine particulate matter (PM) air pollution at 332 locations in the United States. Each file contains data from a single monitor and the ID number for each monitor is contained in the file name. For example, data for monitor 200 is contained in the file "200.csv". 
Pollutantmean function took the two argument (directory name and the indices of monitors whose mean to be found ) and return the mean of the all the monitors given.